+++ src/typings/wp.d.ts
@@ -1,0 +1,2 @@
+// wp.d.ts
+declare module '@wordpress/blocks';