+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 1.10.0
+* Feat: Add Shortcut command (CMD + F).
 * Test: Add e2e tests for plugin codebase.
 * Tested up to WP 6.9.